Add torch_dynamo to resume_in for ease of debugging
resume_in_* code objects show up in user backtraces when failures occur
in code that has been Dynamo processed.  It is obvious to me, a PT2
developer, that these are generated by PT2, but it is NOT obvious to a
non-core dev that this is happened.  Add an extra torch_dynamo
breadcrumb to help get people to the right place.

Signed-off-by: Edward Z. Yang <ezyangmeta.com>

ghstack-source-id: 8f020e007660d45442cab483b534e90831b52653
Pull Request resolved: https://github.com/pytorch/pytorch/pull/118201