[export] Allow constant outputs + None input/outputs (#117894)
Summary:

Added support for constant outputs, which is that we just embed the constant directly into the output, like `return (x, 1)`.
Also adds support for None input/outputs. For None inputs we address it the same way we do to constants, which is that a placeholder with no users will be inserted into the graph, and the None will be embedded into whatever operator is using the None.

Test Plan: CI

Differential Revision: D52881070