[inductor] correctly generate grid info for benchmark_kernel
Previously, we generated the grid argument with tree.numel for
a benchmark TritonKernel. This was not correct, because it
didn't match the launch config used for profiling and running.

This PR fixed the issue by emitting the grid value computed
by the kernel's grid_fn, which is used by the profiler and
the kernel's runner.

ghstack-source-id: eb0a5e23e0ab4c27457909e40d6d4aa5e629abc0
Pull Request resolved: https://github.com/pytorch/pytorch/pull/118202